Suppress error output if Docker processes are already stopped
Before the fix, Docker will output errors if the containers have already been stopped:
```
$ ./stop.sh
Stopping robomaker ... done
Stopping minio     ... done
Removing robomaker ... done
Removing rl_coach  ... done
Removing minio     ... done
Network sagemaker-local is external, skipping
"docker stop" requires at least 1 argument.
See 'docker stop --help'.

Usage:  docker stop [OPTIONS] CONTAINER [CONTAINER...]

Stop one or more running containers
"docker rm" requires at least 1 argument.
See 'docker rm --help'.

Usage:  docker rm [OPTIONS] CONTAINER [CONTAINER...]

Remove one or more containers
```

After the fix, this is output instead:
```
$ ./stop.sh
...
No Docker process running
```

+++ scripts/training/stop.sh
@@ -6,7 +6,15 @@
 # to the commands run from that script, so the export in start.sh does
 # not propagate to here.
 export ROBOMAKER_COMMAND="./run.sh build distributed_training.launch"
+
 docker-compose -f ../../docker/docker-compose.yml down
 
-docker stop $(docker ps | awk ' /sagemaker/ { print $1 }')
-docker rm $(docker ps -a | awk ' /sagemaker/ { print $1 }')
+DOCKER_PROCESS_ID=$(docker ps | awk ' /sagemaker/ { print $1 }')
+
+if [ -z "$DOCKER_PROCESS_ID" ]
+then
+      echo "No Docker process running"
+else
+      docker stop $DOCKER_PROCESS_ID
+      docker rm $(docker ps -a | awk ' /sagemaker/ { print $1 }')
+fi
